refactor : cli 내부 로직 분리 & 디렉토리 구조 변경
cli 내부 main 함수의 로직을 분리하여 cli에서는 전체 흐름만 관리하도록 수정.
cli.js 내부 로직을 setupEnvironment & setupFiles & runTests로 분리
분리 되어 있는 유틸 함수와 로직을 src 내부에서 관리하도록 수정
