+++ src/constants/view.js
