+++ src/utils/formatString.js
@@ -1,5 +1,5 @@
-import {green, red} from "../../utils/consoleColor.js";
-import {RESULT_MSG} from "../../constants/index.js";
+import {green, red} from "./consoleColor.js";
+import {RESULT_MSG} from "../constants/index.js";
 
 export const formatSuccessMessage = (test) => {
   const pathString = test.path === '' ? RESULT_MSG.EMPTY : test.path + RESULT_MSG.DIRECTORY_DELIMITER;
